packages/utss-llm/pyproject.toml
[project]
name = "utss-llm"
version = "0.1.0"
description = "LLM-powered natural language to UTSS strategy parser"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "obichan117" }]
requires-python = ">=3.10"
keywords = [
    "trading",
    "strategy",
    "llm",
    "natural-language",
    "ai",
    "utss",
    "openai",
    "anthropic",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "utss",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
gemini = ["google-generativeai>=0.4.0"]
local = ["ollama>=0.1.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    "ollama>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/obichan117/utss"
Documentation = "https://obichan117.github.io/utss"
Repository = "https://github.com/obichan117/utss"
Issues = "https://github.com/obichan117/utss/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/utss_llm"]

[tool.hatch.build.targets.sdist]
include = ["src/utss_llm"]
